requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
pypdf==4.3.1
pymongo==4.6.3
pytest==8.3.3


